Coerce option & internal improvements
Added the coerce option (:coerce). When set to false, it disables
type coercion. See readme.

Also improved the semantics of the :required option. Rather than checking
if the parameter is nil, it checks the params hash to see if the parameter
is in the request parameters.

Additionally, the fake Boolean "class" (it was a string literal) has been
ontologically substantiated: It is a subclass of TrueClass, patched so
that :=== returns true for instances of TrueClass and FalseClass.